Add parameter to map algebra ingest function for nodata override values
 - closes #306
 - added parameter for skipping pre-processing of source files
 - added parameter for protection level
 - fixed an unrelated failure in quantiles test that randomly failed
 - changed internal handling of nodata values to use double instead of Number
 - wrote common code for processing inputs to be ingested and re-factored the
   command-line ingest and map algebra ingest to use it (IngestInputProcessor)

+++ mrgeo-core/src/test/java/org/mrgeo/data/raster/RasterUtilsTest.java
@@ -153,7 +153,7 @@ public class RasterUtilsTest
 
     scale = 2;
     scaled = RasterUtils.scaleRasterInterp(numberedFloat, 
-        numberedFloat.getWidth() * scale, numberedFloat.getHeight() * scale, Float.NaN);
+        numberedFloat.getWidth() * scale, numberedFloat.getHeight() * scale, Double.NaN);
 
     //    TestUtils.saveRaster(scaled, "TIFF", "scaled.tif");
     //    // scale up
